Move crosspost to share button, make sure only OP can crosspost
